requirements.txt
httpx>=0.25.0
pyyaml>=6.0
flask>=3.0.0
